# perf: faster plugin load