--- a/ape_alchemy/__init__.py
+++ b/ape_alchemy/__init__.py
@@ -1,41 +1,11 @@
 from ape import plugins
 
-from .provider import Alchemy
-
-NETWORKS = {
-    "ethereum": [
-        "mainnet",
-        "sepolia",
-    ],
-    "arbitrum": [
-        "mainnet",
-        "sepolia",
-    ],
-    "base": [
-        "mainnet",
-        "sepolia",
-    ],
-    "fantom": [
-        "opera",
-        "testnet",
-    ],
-    "optimism": [
-        "mainnet",
-        "sepolia",
-    ],
-    "polygon": [
-        "mainnet",
-        "amoy",
-    ],
-    "polygon-zkevm": [
-        "mainnet",
-        "cardona",
-    ],
-}
-
 
 @plugins.register(plugins.ProviderPlugin)
 def providers():
+    from .provider import Alchemy
+    from ._utils import NETWORKS
+
     for ecosystem_name in NETWORKS:
         for network_name in NETWORKS[ecosystem_name]:
             yield ecosystem_name, network_name, Alchemy