--- a/ape_alchemy/__init__.py
+++ b/ape_alchemy/__init__.py
@@ -3,8 +3,8 @@ from ape import plugins
 
 @plugins.register(plugins.ProviderPlugin)
 def providers():
-    from .provider import Alchemy
     from ._utils import NETWORKS
+    from .provider import Alchemy
 
     for ecosystem_name in NETWORKS:
         for network_name in NETWORKS[ecosystem_name]: