--- a/ape_alchemy/__init__.py
+++ b/ape_alchemy/__init__.py
@@ -9,3 +9,17 @@ def providers():
     for ecosystem_name in NETWORKS:
         for network_name in NETWORKS[ecosystem_name]:
             yield ecosystem_name, network_name, Alchemy
+
+
+def __getattr__(name: str):
+    if name == "NETWORKS":
+        from ._utils import NETWORKS
+
+        return NETWORKS
+
+    elif name == "Alchemy":
+        from .provider import Alchemy
+
+        return Alchemy
+
+    raise AttributeError(name)