--- a/ape_alchemy/_utils.py
+++ b/ape_alchemy/_utils.py
@@ -1,0 +1,30 @@
+NETWORKS = {
+    "ethereum": [
+        "mainnet",
+        "sepolia",
+    ],
+    "arbitrum": [
+        "mainnet",
+        "sepolia",
+    ],
+    "base": [
+        "mainnet",
+        "sepolia",
+    ],
+    "fantom": [
+        "opera",
+        "testnet",
+    ],
+    "optimism": [
+        "mainnet",
+        "sepolia",
+    ],
+    "polygon": [
+        "mainnet",
+        "amoy",
+    ],
+    "polygon-zkevm": [
+        "mainnet",
+        "cardona",
+    ],
+}