### What I did

Before:

```
In [1]: %time import ape_alchemy
CPU times: user 1.04 s, sys: 141 ms, total: 1.18 s
Wall time: 1.18 s
```

After:

```
In [1]: %time import ape_alchemy
CPU times: user 7.51 ms, sys: 2.64 ms, total: 10.2 ms
Wall time: 8.89 ms
```

all part of making ape load faster.

### How I did it

### How to verify it

### Checklist

- [ ] Passes all linting checks (pre-commit and CI jobs)
- [ ] New test cases have been added and are passing
- [ ] Documentation has been updated
- [ ] PR title follows [Conventional Commit](https://www.conventionalcommits.org/en/v1.0.0/) standard (will be automatically included in the changelog)